CPM-792: Split big Component and add tests
