+++ components/identifier-generator/back/src/Infrastructure/Symfony/Resources/translations/jsmessages.en_US.yml
