+++ components/identifier-generator/front/src/feature/pages/CreateOrEditGeneratorPage.tsx
@@ -1,7 +1,7 @@
 import React, {useCallback, useState} from 'react';
 import {Button, Helper, TabBar} from 'akeneo-design-system';
 import {PageContent, PageHeader, useTranslate} from '@akeneo-pim-community/shared';
-import {GeneralPropertiesTab, StructureTab} from '../tabs';
+import {GeneralPropertiesTab, Structure} from '../tabs';
 import {IdentifierGenerator} from '../models';
 import {Violation} from '../validators/Violation';
 import {Header} from '../components/Header';
@@ -66,7 +66,7 @@ const CreateOrEditGeneratorPage: React.FC<CreateOrEditGeneratorProps> = ({
             <div>{JSON.stringify(generator.conditions)}</div>
           </>
         )}
-        {currentTab === Tabs.STRUCTURE && <StructureTab />}
+        {currentTab === Tabs.STRUCTURE && <Structure />}
       </PageContent>
     </>
   );
