+++ components/identifier-generator/front/src/feature/tabs/__mocks__/Structure.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Structure: React.FC = () => {
+  return <>StructureMock</>;
+};
+
+export {Structure};
