+++ components/identifier-generator/front/src/feature/tabs/__tests__/Structure.test.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import {render, screen} from '../../tests/test-utils';
+import {Structure} from '../Structure';
+
+jest.mock('../structure/AddPropertyButton');
+
+describe('Structure', () => {
+  it('should render the searchbar', () => {
+    render(<Structure />);
+    expect(screen.getByText('pim_identifier_generator.structure.title')).toBeInTheDocument();
+    expect(screen.getByText('AddPropertyButtonMock')).toBeInTheDocument();
+  });
+});
