+++ components/identifier-generator/front/src/feature/tabs/index.ts
@@ -1,2 +1,2 @@
 export * from './GeneralPropertiesTab';
-export * from './StructureTab';
+export * from './Structure';
