+++ components/identifier-generator/front/src/feature/tabs/structure/__mocks__/AddPropertyButton.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const AddPropertyButton: React.FC = () => {
+  return <>AddPropertyButtonMock</>;
+};
+
+export {AddPropertyButton};
